backend/requirements.txt
Faker==1.0.2
Werkzeug==0.14.1
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.5
django-filter==2.1.0
django-redis==4.10.0
django==2.1.7
djangorestframework==3.9.1
factory-boy==2.11.1
ipython==7.2.0
isort==4.3.4
psycopg2-binary==2.7.4
pytest-django==3.4.7
pytest-sugar==0.9.2
pytest==4.2.1
redis==3.2.0
